backend/requirements.txt
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
python-environ==0.4.54
pytz==2022.6
requests==2.28.1
Werkzeug==2.2.2